API text changes
