--- src/frontend/helpers/types.ts
@@ -1,8 +1,0 @@
-
-type apiEndpointType = {
-	method: string,
-	name: string,
-	url: string,
-	body: any,
-	headers: any,
-}
